refactor: レビュー指摘修正（ユーティリティ外部化・ローディングUI改善・CSS変数統一）
- formatGameDate/formatCreatedAtをlib/format.tsに、getPageNumbersをlib/pagination.tsに切り出し
- ローディング中もヘッダーナビゲーションを表示するよう修正
- GameResultBadgeの関数をRecordマッピングオブジェクトに変更
- エラー表示のハードコードカラーをCSS変数に統一（AdminUsersPageも含む）
